Re-organise IS-05 utility functions into a new class
