+++ IS0501Test.py
@@ -18,14 +18,12 @@
 import requests
 import uuid
 import os
-import re
-import time
 from jsonschema import ValidationError, SchemaError, RefResolver, Draft4Validator
-from random import randint
 
 import TestHelper
 from TestResult import Test
 from GenericTest import GenericTest
+from IS05Utils import IS05Utils
 
 CONN_API_KEY = "connection"
 
@@ -42,15 +40,16 @@ class IS0501Test(GenericTest):
         ]
         GenericTest.__init__(self, apis, omit_paths)
         self.url = self.apis[CONN_API_KEY]["url"]
-        self.senders = self.get_senders()
-        self.receivers = self.get_receivers()
+        self.is05_utils = IS05Utils(self.url)
+        self.senders = self.is05_utils.get_senders()
+        self.receivers = self.is05_utils.get_receivers()
 
     def test_01(self):
         """Api root matches the spec"""
         test = Test("Api root matches the spec")
         expected = ["single/", "bulk/"]
         dest = ""
-        valid, result = self.checkCleanRequestJSON("GET", dest)
+        valid, result = self.is05_utils.checkCleanRequestJSON("GET", dest)
         if valid:
             msg = "Got the wrong json from {} - got {}. Please check json matches the spec, including trailing slashes" \
                 .format(dest, result)
@@ -66,7 +65,7 @@ class IS0501Test(GenericTest):
         test = Test("Single endpoint root matches the spec")
         expected = ["receivers/", "senders/"]
         dest = "single/"
-        valid, result = self.checkCleanRequestJSON("GET", dest)
+        valid, result = self.is05_utils.checkCleanRequestJSON("GET", dest)
         if valid:
             msg = "Got the wrong json from {} - got {}. Please check json matches the spec, including trailing slashes" \
                 .format(dest, result)
@@ -81,7 +80,7 @@ class IS0501Test(GenericTest):
         """Root of /single/senders/ matches the spec"""
         test = Test("Root of /single/senders/ matches the spec")
         dest = "single/senders/"
-        valid, response = self.checkCleanRequestJSON("GET", dest)
+        valid, response = self.is05_utils.checkCleanRequestJSON("GET", dest)
         smsg = "UUIDs missing trailing slashes in response from {}".format(dest)
         umsg = "Response from {} containts invalid UUIDs".format(dest)
         amsg = "Expected an array from {}, got {}".format(dest, type(response))
@@ -110,7 +109,7 @@ class IS0501Test(GenericTest):
         """Root of /single/receivers/ matches the spec"""
         test = Test("Root of /single/receivers/ matches the spec")
         dest = "single/receivers/"
-        valid, response = self.checkCleanRequestJSON("GET", dest)
+        valid, response = self.is05_utils.checkCleanRequestJSON("GET", dest)
         smsg = "UUIDs missing trailing slashes in response from {}".format(dest)
         umsg = "Response from {} containts invalid UUIDs".format(dest)
         amsg = "Expected an array from {}, got {}".format(dest, type(response))
@@ -141,7 +140,7 @@ class IS0501Test(GenericTest):
         if len(self.senders) > 0:
             for sender in self.senders:
                 dest = "single/senders/" + sender + "/"
-                valid, response = self.checkCleanRequestJSON("GET", dest)
+                valid, response = self.is05_utils.checkCleanRequestJSON("GET", dest)
                 expected = [
                     "constraints/",
                     "staged/",
@@ -166,7 +165,7 @@ class IS0501Test(GenericTest):
         if len(self.receivers) > 0:
             for receiver in self.receivers:
                 dest = "single/receivers/" + receiver + "/"
-                valid, response = self.checkCleanRequestJSON("GET", dest)
+                valid, response = self.is05_utils.checkCleanRequestJSON("GET", dest)
                 expected = [
                     "constraints/",
                     "staged/",
@@ -220,7 +219,7 @@ class IS0501Test(GenericTest):
         """All params listed in /single/senders/<uuid>/constraints/ matches /staged/ and /active/"""
         test = Test("All params listed in /single/senders/<uuid>/constraints/ matches /staged/ and /active/")
         if len(self.senders) > 0:
-            valid, response = self.check_params_match("senders", self.senders)
+            valid, response = self.is05_utils.check_params_match("senders", self.senders)
             if valid:
                 return test.PASS()
             else:
@@ -235,7 +234,7 @@ class IS0501Test(GenericTest):
         """All params listed in /single/receivers/<uuid>/constraints/ matches /staged/ and /active/"""
         test = Test("All params listed in /single/receivers/<uuid>/constraints/ matches /staged/ and /active/")
         if len(self.receivers) > 0:
-            valid, response = self.check_params_match("receivers", self.receivers)
+            valid, response = self.is05_utils.check_params_match("receivers", self.receivers)
             if valid:
                 return test.PASS()
             else:
@@ -264,7 +263,7 @@ class IS0501Test(GenericTest):
             for sender in self.senders:
                 dest = "single/senders/" + sender + "/constraints/"
                 try:
-                    valid, response = self.checkCleanRequestJSON("GET", dest)
+                    valid, response = self.is05_utils.checkCleanRequestJSON("GET", dest)
                     if valid:
                         if len(response) > 0 and isinstance(response[0], dict):
                             params = response[0].keys()
@@ -306,7 +305,7 @@ class IS0501Test(GenericTest):
             for receiver in self.receivers:
                 dest = "single/receivers/" + receiver + "/constraints/"
                 try:
-                    valid, response = self.checkCleanRequestJSON("GET", dest)
+                    valid, response = self.is05_utils.checkCleanRequestJSON("GET", dest)
                     if valid:
                         if len(response) > 0 and isinstance(response[0], dict):
                             params = response[0].keys()
@@ -416,7 +415,7 @@ class IS0501Test(GenericTest):
         """Sender invalid patch is refused"""
         test = Test("Sender invalid patch is refused")
         if len(self.senders) > 0:
-            valid, response = self.check_refuses_invalid_patch("sender", self.senders)
+            valid, response = self.is05_utils.check_refuses_invalid_patch("sender", self.senders)
             if valid:
                 return test.PASS()
             else:
@@ -428,7 +427,7 @@ class IS0501Test(GenericTest):
         """Receiver invalid patch is refused"""
         test = Test("Receiver invalid patch is refused")
         if len(self.receivers) > 0:
-            valid, response = self.check_refuses_invalid_patch("receiver", self.receivers)
+            valid, response = self.is05_utils.check_refuses_invalid_patch("receiver", self.receivers)
             if valid:
                 return test.PASS()
             else:
@@ -444,9 +443,9 @@ class IS0501Test(GenericTest):
                 url = "single/receivers/" + receiver + "/staged"
                 id = str(uuid.uuid4())
                 data = {"sender_id": id}
-                valid, response = self.checkCleanRequestJSON("PATCH", url, data=data)
+                valid, response = self.is05_utils.checkCleanRequestJSON("PATCH", url, data=data)
                 if valid:
-                    valid2, response2 = self.checkCleanRequestJSON("GET", url + "/")
+                    valid2, response2 = self.is05_utils.checkCleanRequestJSON("GET", url + "/")
                     if valid2:
                         try:
                             senderId = response['sender_id']
@@ -473,9 +472,9 @@ class IS0501Test(GenericTest):
                 url = "single/senders/" + sender + "/staged"
                 id = str(uuid.uuid4())
                 data = {"receiver_id": id}
-                valid, response = self.checkCleanRequestJSON("PATCH", url, data=data)
+                valid, response = self.is05_utils.checkCleanRequestJSON("PATCH", url, data=data)
                 if valid:
-                    valid2, response2 = self.checkCleanRequestJSON("GET", url + "/")
+                    valid2, response2 = self.is05_utils.checkCleanRequestJSON("GET", url + "/")
                     if valid2:
                         try:
                             receiverId = response['receiver_id']
@@ -499,10 +498,10 @@ class IS0501Test(GenericTest):
         test = Test("Sender transport parameters are changeable")
         if len(self.senders) > 0:
             for sender in self.senders:
-                valid, values = self.generate_destination_ports("sender", sender)
+                valid, values = self.is05_utils.generate_destination_ports("sender", sender)
                 if valid:
-                    valid2, response2 = self.check_change_transport_param("sender", self.senders,
-                                                                          "destination_port", values, sender)
+                    valid2, response2 = self.is05_utils.check_change_transport_param("sender", self.senders,
+                                                                                     "destination_port", values, sender)
                     if valid2:
                         pass
                     else:
@@ -518,10 +517,11 @@ class IS0501Test(GenericTest):
         test = Test("Receiver transport parameters are changeable")
         if len(self.receivers) > 0:
             for receiver in self.receivers:
-                valid, values = self.generate_destination_ports("receiver", receiver)
+                valid, values = self.is05_utils.generate_destination_ports("receiver", receiver)
                 if valid:
-                    valid2, response2 = self.check_change_transport_param("receiver", self.receivers,
-                                                                          "destination_port", values, receiver)
+                    valid2, response2 = self.is05_utils.check_change_transport_param("receiver", self.receivers,
+                                                                                     "destination_port", values,
+                                                                                     receiver)
                     if valid2:
                         pass
                     else:
@@ -537,7 +537,8 @@ class IS0501Test(GenericTest):
         test = Test("Immediate activation of a sender is possible")
         if len(self.senders) > 0:
             for sender in self.senders:
-                valid, response = self.check_activation("sender", sender, self.check_perform_immediate_activation)
+                valid, response = self.is05_utils.check_activation("sender", sender,
+                                                                   self.is05_utils.check_perform_immediate_activation)
                 if valid:
                     pass
                 else:
@@ -551,7 +552,8 @@ class IS0501Test(GenericTest):
         test = Test("Immediate activation of a receiver is possible")
         if len(self.receivers) > 0:
             for receiver in self.receivers:
-                valid, response = self.check_activation("receiver", receiver, self.check_perform_immediate_activation)
+                valid, response = self.is05_utils.check_activation("receiver", receiver,
+                                                                   self.is05_utils.check_perform_immediate_activation)
                 if valid:
                     pass
                 else:
@@ -565,7 +567,8 @@ class IS0501Test(GenericTest):
         test = Test("Relative activation of a sender is possible")
         if len(self.senders) > 0:
             for sender in self.senders:
-                valid, response = self.check_activation("sender", sender, self.check_perform_relative_activation)
+                valid, response = self.is05_utils.check_activation("sender", sender,
+                                                                   self.is05_utils.check_perform_relative_activation)
                 if valid:
                     pass
                 else:
@@ -579,8 +582,8 @@ class IS0501Test(GenericTest):
         test = Test("Relative activation of a receiver is possible")
         if len(self.receivers) > 0:
             for receiver in self.receivers:
-                valid, response = self.check_activation("receiver", receiver,
-                                                        self.check_perform_relative_activation)
+                valid, response = self.is05_utils.check_activation("receiver", receiver,
+                                                                   self.is05_utils.check_perform_relative_activation)
                 if valid:
                     pass
                 else:
@@ -594,7 +597,8 @@ class IS0501Test(GenericTest):
         test = Test("Absolute activation of a sender is possible")
         if len(self.senders) > 0:
             for sender in self.senders:
-                valid, response = self.check_activation("sender", sender, self.check_perform_absolute_activation)
+                valid, response = self.is05_utils.check_activation("sender", sender,
+                                                                   self.is05_utils.check_perform_absolute_activation)
                 if valid:
                     pass
                 else:
@@ -608,8 +612,8 @@ class IS0501Test(GenericTest):
         test = Test("Absolute activation of a receiver is possible")
         if len(self.receivers) > 0:
             for receiver in self.receivers:
-                valid, response = self.check_activation("receiver", receiver,
-                                                        self.check_perform_absolute_activation)
+                valid, response = self.is05_utils.check_activation("receiver", receiver,
+                                                                   self.is05_utils.check_perform_absolute_activation)
                 if valid:
                     pass
                 else:
@@ -654,7 +658,7 @@ class IS0501Test(GenericTest):
         """/bulk/ endpoint returns correct JSON"""
         test = Test("/bulk/ endpoint returns correct JSON")
         url = "bulk/"
-        valid, response = self.checkCleanRequestJSON("GET", url)
+        valid, response = self.is05_utils.checkCleanRequestJSON("GET", url)
         if valid:
             expected = ['senders/', 'receivers/']
             msg = "Got wrong response from {}, expected an array containing {}, got {}".format(url, expected, response)
@@ -669,7 +673,7 @@ class IS0501Test(GenericTest):
         """GET on /bulk/senders returns 405"""
         test = Test("GET on /bulk/senders returns 405")
         url = "bulk/senders"
-        valid, response = self.checkCleanRequestJSON("GET", url, code=405)
+        valid, response = self.is05_utils.checkCleanRequestJSON("GET", url, code=405)
         if valid:
             return test.PASS()
         else:
@@ -679,7 +683,7 @@ class IS0501Test(GenericTest):
         """GET on /bulk/receivers returns 405"""
         test = Test("GET on /bulk/receivers returns 405")
         url = "bulk/receivers"
-        valid, response = self.checkCleanRequestJSON("GET", url, code=405)
+        valid, response = self.is05_utils.checkCleanRequestJSON("GET", url, code=405)
         if valid:
             return test.PASS()
         else:
@@ -715,7 +719,7 @@ class IS0501Test(GenericTest):
         if len(self.senders) > 0:
             for sender in self.senders:
                 url = "single/senders/{}/".format(sender)
-                valid, response = self.check_num_legs(url, "sender", sender)
+                valid, response = self.is05_utils.check_num_legs(url, "sender", sender)
                 if valid:
                     pass
                 else:
@@ -730,7 +734,7 @@ class IS0501Test(GenericTest):
         if len(self.receivers) > 0:
             for receiver in self.receivers:
                 url = "single/receivers/{}/".format(receiver)
-                valid, response = self.check_num_legs(url, "receiver", receiver)
+                valid, response = self.is05_utils.check_num_legs(url, "receiver", receiver)
                 if valid:
                     pass
                 else:
@@ -739,72 +743,13 @@ class IS0501Test(GenericTest):
         else:
             return test.NA("Not tested. No resources found.")
 
-    def check_num_legs(self, url, type, uuid):
-        """Checks the number of legs present on a given sender/receiver"""
-        max = 2
-        min = 1
-        constraintsUrl = url + "constraints/"
-        stagedUrl = url + "staged/"
-        activeUrl = url + "active/"
-        valid1, constraints = self.checkCleanRequestJSON("GET", constraintsUrl)
-        if valid1:
-            valid2, staged = self.checkCleanRequestJSON("GET", stagedUrl)
-            if valid2:
-                valid3, active = self.checkCleanRequestJSON("GET", activeUrl)
-                if valid3:
-                    try:
-                        stagedParams = staged['transport_params']
-                    except TypeError:
-                        return False, "Expected an object to be returned from {}, got {}: {}".format(stagedUrl,
-                                                                                                     type(staged),
-                                                                                                     staged)
-                    except KeyError:
-                        return False, "Could not find transport params in object from {}, got {}".format(stagedUrl,
-                                                                                                         staged)
-                    try:
-                        activeParams = active['transport_params']
-                    except TypeError:
-                        return False, "Expected an object to be returned from {}, got {}: {}".format(activeUrl,
-                                                                                                     type(active),
-                                                                                                     active)
-                    except KeyError:
-                        return False, "Could not find transport params in object from {}, got {}".format(activeUrl,
-                                                                                                         active)
-                    if len(constraints) <= max:
-                        pass
-                    else:
-                        return False, "{} {} has too many legs".format(type, uuid)
-                    if len(constraints) >= min:
-                        pass
-                    else:
-                        return False, "{} {} has too few legs".format(type, uuid)
-                    if len(constraints) == len(stagedParams):
-                        pass
-                    else:
-                        return False, "Number of legs in constraints and staged is different for {} {}".format(type,
-                                                                                                               uuid)
-
-                    if len(constraints) == len(activeParams):
-                        pass
-                    else:
-                        return False, "Number of legs in constraints and active is different for {} {}".format(type,
-                                                                                                               uuid)
-
-                    return True, ""
-                else:
-                    return False, active
-            else:
-                return False, staged
-        else:
-            return False, constraints
-
     def check_bulk_stage(self, port, portList):
         """Test changing staged parameters on the bulk interface"""
         url = self.url + "bulk/" + port + "s"
         data = []
         ports = {}
         for portInst in portList:
-            valid, response = self.generate_destination_ports(port, portInst)
+            valid, response = self.is05_utils.generate_destination_ports(port, portInst)
             if valid:
                 ports[portInst] = response
                 toAdd = {}
@@ -838,9 +783,9 @@ class IS0501Test(GenericTest):
         for portInst in portList:
             activeUrl = "single/" + port + "s/" + portInst + "/staged/"
 
-            valid, response = self.checkCleanRequestJSON("GET", activeUrl)
+            valid, response = self.is05_utils.checkCleanRequestJSON("GET", activeUrl)
             if valid:
-                for i in range(0, self.get_num_paths(portInst, port)):
+                for i in range(0, self.is05_utils.get_num_paths(portInst, port)):
                     try:
                         value = response['transport_params'][i]['destination_port']
                     except KeyError:
@@ -858,396 +803,12 @@ class IS0501Test(GenericTest):
                 return False, response
         return True, ""
 
-    def check_staged_activation_params_default(self, port, portId):
-        # Check that the staged activation parameters have returned to their default values
-        stagedUrl = "single/" + port + "s/" + portId + "/staged"
-        valid, response = self.checkCleanRequestJSON("GET", stagedUrl)
-        if valid:
-            expected = {"mode": None, "requested_time": None, "activation_time": None}
-            try:
-                params = response['activation']
-            except KeyError:
-                return False, "Could not find a receiver_id entry in response from {}, got {}".format(stagedUrl,
-                                                                                                      response)
-            except TypeError:
-                return False, "Expected a dict to be returned from {}, got a {}: {}".format(stagedUrl, type(response),
-                                                                                            response)
-            if params == expected:
-                return True, ""
-            else:
-                msg = "Activation parameters in staged have not returned to their defaults after activation. " \
-                      "Expected {}, got {}".format(expected, params)
-                return False, msg
-        else:
-            return False, response
-
-    def check_perform_immediate_activation(self, port, portId, stagedParams):
-        # Request an immediate activation
-        stagedUrl = "single/" + port + "s/" + portId + "/staged"
-        activeUrl = "single/" + port + "s/" + portId + "/active"
-        data = {"activation": {"mode": "activate_immediate"}}
-        valid, response = self.checkCleanRequestJSON("PATCH", stagedUrl, data=data)
-        if valid:
-            try:
-                mode = response['activation']['mode']
-                requested = response['activation']['requested_time']
-                activation = response['activation']['activation_time']
-            except KeyError:
-                return False, "Could not find all activation entries from {}, got {}".format(stagedUrl, response)
-            except TypeError:
-                return False, "Expected a dict to be returned from {}, got a {}: {}".format(stagedUrl, type(response),
-                                                                                            response)
-            mmsg = "Unexpected mode returned: expected `activate_immediate`, got {}".format(mode)
-            rmsg = "Expected null requested time for immediate activation, got {}".format(requested)
-            amsg = "Expected an activation time matching the regex ^[0-9]+:[0-9]+$, but got {}".format(activation)
-            if mode == "activate_immediate":
-                pass
-            else:
-                return False, mmsg
-            if requested is None:
-                pass
-            else:
-                return False, rmsg
-            try:
-                if re.match("^[0-9]+:[0-9]+$", activation) is not None:
-                    pass
-                else:
-                    return False, amsg
-            except TypeError:
-                return False, amsg
-
-            valid2, response2 = self.check_staged_activation_params_default(port, portId)
-            if valid2:
-                # Check the values now on /active
-                valid3, response3 = self.checkCleanRequestJSON("GET", activeUrl)
-                if valid3:
-                    for i in range(0, self.get_num_paths(portId, port)):
-                        try:
-                            activePort = response3['transport_params'][i]['destination_port']
-                        except KeyError:
-                            return False, "Could not find active destination_port entry on leg {} from {}, " \
-                                          "got {}".format(i, activeUrl, response3)
-                        except TypeError:
-                            return False, "Expected a dict to be returned from {} on leg {}, got a {}: {}".format(
-                                activeUrl, i, type(response3), response3)
-                        try:
-                            stagedPort = stagedParams[i]['destination_port']
-                        except KeyError:
-                            return False, "Could not find staged destination_port entry on leg {} from {}, " \
-                                          "got {}".format(i, stagedUrl, stagedParams)
-                        except TypeError:
-                            return False, "Expected a dict to be returned from {} on leg {}, got a {}: {}".format(
-                                stagedUrl, i, type(response3), stagedParams)
-                        msg = "Transport parameters did not transition to active during an imeddiate activation"
-                        if stagedPort == activePort:
-                            pass
-                        else:
-                            return False, msg
-
-                    msg = "Activation mode was not set to `activate_immediate` at {} " \
-                          "after an immediate activation".format(activeUrl)
-                    if response3['activation']['mode'] == "activate_immediate":
-                        pass
-                    else:
-                        return False, msg
-                else:
-                    return False, response3
-            else:
-                return False, response2
-        else:
-            return False, response
-        return True, ""
-
-    def check_perform_relative_activation(self, port, portId, stagedParams):
-        # Request an relative activation
-        stagedUrl = "single/" + port + "s/" + portId + "/staged"
-        activeUrl = "single/" + port + "s/" + portId + "/active"
-        data = {"activation": {"mode": "activate_scheduled_relative", "requested_time": "0:2"}}
-        valid, response = self.checkCleanRequestJSON("PATCH", stagedUrl, data=data, code=202)
-        if valid:
-            try:
-                mode = response['activation']['mode']
-                requested = response['activation']['requested_time']
-                activation = response['activation']['activation_time']
-            except KeyError:
-                return False, "Could not find all activation entries from {}, got {}".format(stagedUrl, response)
-            except TypeError:
-                return False, "Expected a dict to be returned from {}, got a {}: {}".format(stagedUrl, type(response),
-                                                                                            response)
-            mmsg = "Expected mode `activate_sechduled_relative` for relative activation, got {}".format(mode)
-            rmsg = "Expected requested time `0:2` for relative activation, got {}".format(requested)
-            amsg = "Expected activation time to match regex ^[0-9]+:[0-9]+$, got {}".format(activation)
-            if mode == "activate_scheduled_relative":
-                pass
-            else:
-                return False, mmsg
-            if requested == "0:2":
-                pass
-            else:
-                return False, rmsg
-            if re.match("^[0-9]+:[0-9]+$", activation) is not None:
-                pass
-            else:
-                return False, amsg
-            time.sleep(0.2)
-
-            retries = 0
-            finished = False
-
-            while retries < 5 and not finished:
-                # Check the values now on /active
-                valid2, activeParams = self.checkCleanRequestJSON("GET", activeUrl)
-                if valid2:
-                    for i in range(0, self.get_num_paths(portId, port)):
-                        try:
-                            activePort = activeParams['transport_params'][i]['destination_port']
-                        except KeyError:
-                            return False, "Could not find active destination_port entry on leg {} from {}, " \
-                                          "got {}".format(i, activeUrl, activeParams)
-                        except TypeError:
-                            return False, "Expected a dict to be returned from {} on leg {}, " \
-                                          "got a {}: {}".format(activeUrl, i, type(activeParams), activeParams)
-                        try:
-                            stagedPort = stagedParams[i]['destination_port']
-                        except KeyError:
-                            return False, "Could not find staged destination_port entry on leg {} from {}, " \
-                                          "got {}".format(i, stagedUrl, stagedParams)
-                        except TypeError:
-                            return False, "Expected a dict to be returned from {} on leg {}, " \
-                                          "got a {}: {}".format(stagedUrl, i, type(activeParams), stagedParams)
-                        if stagedPort == activePort:
-                            finished = True
-                        else:
-                            retries = retries + 1
-                            time.sleep(0.2)
-
-                    if finished:
-                        try:
-                            if activeParams['activation']['mode'] == "activate_scheduled_relative":
-                                if retries > 0:
-                                    return True, "(Retries: {})".format(str(retries))
-                                else:
-                                    return True, ""
-                            else:
-                                return False, "Activation mode was not set to `activate_scheduled_relative` at {} " \
-                                              "after a relative activation".format(activeUrl)
-                        except KeyError:
-                            return False, "Expected 'mode' key in 'activation' object."
-
-                else:
-                    return False, activeParams
-            return False, "Transport parameters did not transition to active during an relative activation " \
-                          "(Retries: {})".format(str(retries))
-        else:
-            return False, response
-
-    def check_perform_absolute_activation(self, port, portId, stagedParams):
-        # request an absolute activation
-        stagedUrl = "single/" + port + "s/" + portId + "/staged"
-        activeUrl = "single/" + port + "s/" + portId + "/active"
-        TAItime = TestHelper.getTAITime(1)
-        data = {"activation": {"mode": "activate_scheduled_absolute", "requested_time": TAItime}}
-        valid, response = self.checkCleanRequestJSON("PATCH", stagedUrl, data=data, code=202)
-        if valid:
-            try:
-                mode = response['activation']['mode']
-                requested = response['activation']['requested_time']
-                activation = response['activation']['activation_time']
-            except KeyError:
-                return False, "Could not find all activation entries from {}, got {}".format(stagedUrl, response)
-            except TypeError:
-                return False, "Expected a dict to be returned from {}, got a {}: {}".format(stagedUrl, type(response),
-                                                                                            response)
-            mmsg = "Expected mode `activate_sechduled_absolute` for relative activation, got {}".format(mode)
-            rmsg = "Expected requested time `{}` for relative activation, got {}".format(TAItime, requested)
-            amsg = "Expected activation time to match regex ^[0-9]+:[0-9]+$, got {}".format(activation)
-            try:
-                if response['activation']['mode'] == "activate_scheduled_absolute":
-                    pass
-                else:
-                    return False, mmsg
-            except KeyError:
-                return False, "Expected 'mode' key in 'activation' object."
-            try:
-                if response['activation']['requested_time'] == TAItime:
-                    pass
-                else:
-                    return False, rmsg
-            except KeyError:
-                return False, "Expected 'requested_time' key in 'activation' object."
-            try:
-                if re.match("^[0-9]+:[0-9]+$", response['activation']['activation_time']) is not None:
-                    pass
-                else:
-                    return False, amsg
-            except KeyError:
-                return False, "Expected 'activation_time' key in 'activation' object."
-            # Allow extra time for processing between getting time and making request
-            time.sleep(2)
-
-            retries = 0
-            finished = False
-
-            while retries < 5 and not finished:
-                # Check the values now on /active
-                valid2, activeParams = self.checkCleanRequestJSON("GET", activeUrl)
-                if valid2:
-                    for i in range(0, self.get_num_paths(portId, port)):
-                        try:
-                            activePort = activeParams['transport_params'][i]['destination_port']
-                        except KeyError:
-                            return False, "Could not find active destination_port entry on leg {} from {}, " \
-                                          "got {}".format(i, activeUrl, activeParams)
-                        except TypeError:
-                            return False, "Expected a dict to be returned from {} on leg {}, got a {}: " \
-                                          "{}".format(activeUrl, i, type(activeParams), activeParams)
-                        try:
-                            stagedPort = stagedParams[i]['destination_port']
-                        except KeyError:
-                            return False, "Could not find staged destination_port entry on leg {} from {}, " \
-                                          "got {}".format(i, stagedUrl, stagedParams)
-                        except TypeError:
-                            return False, "Expected a dict to be returned from {} on leg {}, got a {}: " \
-                                          "{}".format(stagedUrl, i, type(activeParams), stagedParams)
-                        if activePort == stagedPort:
-                            finished = True
-                        else:
-                            retries = retries + 1
-                            time.sleep(1)
-
-                    if finished:
-                        try:
-                            if activeParams['activation']['mode'] == "activate_scheduled_absolute":
-                                if retries > 0:
-                                    return True, "(Retries: {})".format(str(retries))
-                                else:
-                                    return True, ""
-                            else:
-                                return False, "Activation mode was not set to `activate_scheduled_absolute` at {} " \
-                                              "after a absolute activation".format(activeUrl)
-                        except KeyError:
-                            return False, "Expected 'mode' key in 'activation' object."
-                else:
-                    return False, activeParams
-            return False, "Transport parameters did not transition to active during an absolute activation " \
-                          "(Retries: {})".format(str(retries))
-        else:
-            return False, response
-
-    def check_activation(self, port, portId, activationMethod):
-        """Checks that when an immediate activation is called staged parameters are moved
-        to active and the activation is correctly displayed in the /active endpoint"""
-        # Set a new destination port in staged
-        valid, destinationPort = self.generate_destination_ports(port, portId)
-        if valid:
-            stagedUrl = "single/" + port + "s/" + portId + "/staged"
-            data = {"transport_params": []}
-            for i in range(0, self.get_num_paths(portId, port)):
-                data['transport_params'].append({"destination_port": destinationPort[i]})
-            valid2, r = self.checkCleanRequestJSON("PATCH", stagedUrl, data=data)
-            if valid2:
-                try:
-                    stagedParams = r['transport_params']
-                except KeyError:
-                    return False, "Could not find `transport_params` entry in response from {}".format(stagedUrl)
-                except TypeError:
-                    return False, "Expected a dict to be returned from {}, got a {}".format(stagedUrl,
-                                                                                            type(stagedParams))
-                return activationMethod(port, portId, stagedParams)
-            else:
-                return False, r
-        else:
-            return False, destinationPort
-
-    def generate_destination_ports(self, port, portId):
-        """Uses a port's constraints to generate an allowable destination
-        ports for it"""
-        url = "single/" + port + "s/" + portId + "/constraints/"
-        valid, constraints = self.checkCleanRequestJSON("GET", url)
-        if valid:
-            toReturn = []
-            try:
-                for entry in constraints:
-                    if "enum" in entry['destination_port']:
-                        values = entry['destination_port']['enum']
-                        toReturn.append(values[randint(0, len(values) - 1)])
-                    else:
-                        if "minimum" in entry['destination_port']:
-                            min = entry['destination_port']['minimum']
-                        else:
-                            min = 5000
-                        if "maximum" in entry['destination_port']:
-                            max = entry['destination_port']['maximum']
-                        else:
-                            max = 49151
-                        toReturn.append(randint(min, max))
-                return True, toReturn
-            except TypeError:
-                return False, "Expected a dict to be returned from {}, got a {}: {}".format(url, type(constraints),
-                                                                                            constraints)
-        else:
-            return False, constraints
-
-    def check_change_transport_param(self, port, portList, paramName, paramValues, myPort):
-        """Check that we can update a transport parameter"""
-        url = "single/" + port + "s/" + myPort + "/staged"
-        data = {}
-        data['transport_params'] = []
-        paths = self.get_num_paths(myPort, port)
-        for i in range(0, paths):
-            data['transport_params'].append({})
-            data['transport_params'][i][paramName] = paramValues[i]
-        valid, response = self.checkCleanRequestJSON("PATCH", url, data=data)
-        if valid:
-            valid2, response2 = self.checkCleanRequestJSON("GET", url + "/")
-            if valid2:
-                try:
-                    response3 = response2['transport_params']
-                except KeyError:
-                    return False, "Could not find transport_params in response from {}, got {}".format(url, response2)
-                except TypeError:
-                    return False, "Expected a dict to be returned from {}, got a {}: {}".format(url, type(response2),
-                                                                                                response2)
-                count = 0
-                try:
-                    for item in response3:
-                        expected = paramValues[count]
-                        actual = item[paramName]
-                        msg = "Could not change {} parameter at {}, expected {}, got {}".format(paramName, url,
-                                                                                                expected,
-                                                                                                actual)
-                        if actual == expected:
-                            pass
-                        else:
-                            return False, msg
-                        count = count + 1
-                except TypeError:
-                    return False, "Expected a dict to be returned from {}, got a {}: {}".format(url, type(response3),
-                                                                                                response3)
-            else:
-                return False, response2
-        else:
-            return False, response
-        return True, ""
-
-    def check_refuses_invalid_patch(self, port, portList):
-        """Check that invalid patch requests to /staged are met with an HTTP 400"""
-        data = {"bad": "data"}
-        for myPort in portList:
-            url = "single/" + port + "s/" + myPort + "/staged"
-            valid, response = self.checkCleanRequestJSON("PATCH", url, data=data, code=400)
-            if valid:
-                pass
-            else:
-                return False, response
-        return True, ""
-
     def check_patch_response_schema_valid(self, port, portList):
         """Check the response to an empty patch request complies with the schema"""
         for myPort in portList:
             url = "single/" + port + "s/" + myPort + "/staged"
             data = {}
-            valid, response = self.checkCleanRequestJSON("PATCH", url, data=data)
+            valid, response = self.is05_utils.checkCleanRequestJSON("PATCH", url, data=data)
             if valid:
                 schema = self.get_schema(CONN_API_KEY, "PATCH", "/single/" + port + "s/{" + port + "Id}/staged", 200)
                 try:
@@ -1263,14 +824,15 @@ class IS0501Test(GenericTest):
         the constents of the /constraints endpoint"""
         for myPort in portList:
             dest = "single/" + port + "s/" + myPort + "/staged/"
-            valid, response = self.checkCleanRequestJSON("GET", dest)
+            valid, response = self.is05_utils.checkCleanRequestJSON("GET", dest)
             if valid:
                 schema = self.load_schema(CONN_API_KEY, "v1.0_" + port + "_transport_params_rtp.json")
                 resolver = RefResolver(self.file_prefix + os.path.join(self.apis[CONN_API_KEY]["spec_path"] +
                                                                        '/APIs/schemas/'),
                                        schema)
-                constraints_valid, constraints_response = self.checkCleanRequestJSON("GET", "single/" + port + "s/" +
-                                                                                     myPort + "/constraints/")
+                constraints_valid, constraints_response = self.is05_utils.checkCleanRequestJSON("GET", "single/" +
+                                                                                                port + "s/" + myPort +
+                                                                                                "/constraints/")
                 if constraints_valid:
                     count = 0
                     try:
@@ -1294,148 +856,11 @@ class IS0501Test(GenericTest):
                 return False, response
         return True, ""
 
-    def check_params_match(self, port, portList):
-        """Generic test for checking params listed in the /constraints endpoint
-        are listed in in the /staged and /active endpoints"""
-        for myPort in portList:
-            rDest = "single/" + port + "/" + myPort + "/constraints/"
-            sDest = "single/" + port + "/" + myPort + "/staged/"
-            aDest = "single/" + port + "/" + myPort + "/active/"
-            r_valid, r_response = self.checkCleanRequestJSON("GET", rDest)
-            s_valid, s_response = self.checkCleanRequestJSON("GET", sDest)
-            a_valid, a_response = self.checkCleanRequestJSON("GET", aDest)
-            count = 0
-            amsg = "Expected an array to be returned {} but got {}".format(rDest, r_response)
-            omsg = "Expected array entries to be dictionaries at {} but got {}".format(rDest, r_response)
-            # Check the response is a list
-            if r_valid:
-                if s_valid:
-                    if a_valid:
-                        if isinstance(r_response, list):
-                            if len(r_response) > 0:
-                                for entry in r_response:
-                                    if isinstance(entry, dict):
-                                        constraintParams = entry.keys()
-                                        try:
-                                            stagedParams = s_response['transport_params'][count].keys()
-                                        except AttributeError:
-                                            # Found something that we couldn't get keys from, not a dict then...
-                                            return False, "Staged parameters contain non-dicts in array position " \
-                                                          "{}".format(count)
-                                        except KeyError:
-                                            return False, "Staged parameters do not contain transport_params"
-                                        try:
-                                            activeParams = a_response['transport_params'][count].keys()
-                                        except AttributeError:
-                                            # Found something that we couldn't get keys from, not a dict then...
-                                            return False, "Active parameters contain non-dicts in array position " \
-                                                          "{}".format(count)
-                                        except KeyError:
-                                            return False, "Active parameters do not contain transport_params"
-                                        smsg = "Staged parameter set does not match parameters in constraints"
-                                        amsg = "Active parameter set does not match parameters in constraints"
-                                        if len(r_response) == len(s_response['transport_params']):
-                                            pass
-                                        else:
-                                            return False, "Number of legs differs between staged and constraints"
-
-                                        if len(r_response) == len(a_response['transport_params']):
-                                            pass
-                                        else:
-                                            return False, "Number of legs differs between active and constraints"
-
-                                        if TestHelper.compare_json(constraintParams, stagedParams):
-                                            pass
-                                        else:
-                                            return False, smsg
-
-                                        if TestHelper.compare_json(constraintParams, activeParams):
-                                            pass
-                                        else:
-                                            return False, amsg
-                                        count = count + 1
-                                    else:
-                                        return False, omsg
-                            else:
-                                return False, "Not tested. No resources found."
-                        else:
-                            return False, amsg
-                    else:
-                        return False, a_response
-                else:
-                    return False, s_response
-            else:
-                return False, r_response
-        return True, ""
-
-    def get_senders(self):
-        """Gets a list of the available senders on the API"""
-        toReturn = []
-        try:
-            r = requests.get(self.url + "single/senders/")
-            try:
-                for value in r.json():
-                    toReturn.append(value[:-1])
-            except ValueError:
-                pass
-        except requests.exceptions.RequestException:
-            pass
-        return toReturn
-
-    def get_receivers(self):
-        """Gets a list of the available receivers on the API"""
-        toReturn = []
-        try:
-            r = requests.get(self.url + "single/receivers/")
-            try:
-                for value in r.json():
-                    toReturn.append(value[:-1])
-            except ValueError:
-                pass
-        except requests.exceptions.RequestException:
-            pass
-        return toReturn
-
-    def get_num_paths(self, port, portType):
-        """Returns the number or redundant paths on a port"""
-        url = self.url + "single/" + portType + "s/" + port + "/constraints/"
-        try:
-            r = requests.get(url)
-            try:
-                rjson = r.json()
-                return len(rjson)
-            except ValueError:
-                return 0
-        except requests.exceptions.RequestException:
-            return 0
 
     def compare_to_schema(self, schema, endpoint, status_code=200):
         """Compares the response from an endpoint to a schema"""
-        valid, response = self.checkCleanRequest("GET", endpoint, code=status_code)
+        valid, response = self.is05_utils.checkCleanRequest("GET", endpoint, code=status_code)
         if valid:
             return self.check_response(CONN_API_KEY, schema, "GET", response)
         else:
             return False, "Invalid response while getting data: " + response
-
-    def checkCleanRequest(self, method, dest, data=None, code=200):
-        """Checks a request can be made and the resulting json can be parsed"""
-        status, response = self.do_request(method, self.url + dest, data)
-        if not status:
-            return status, response
-
-        message = "Expected status code {} from {}, got {}.".format(code, dest, response.status_code)
-        if response.status_code == code:
-            return True, response
-        else:
-            return False, message
-
-    def checkCleanRequestJSON(self, method, dest, data=None, code=200):
-        valid, response = self.checkCleanRequest(method, dest, data, code)
-        if valid:
-            try:
-                return True, response.json()
-            except:
-                # Failed parsing JSON
-                return False, "Invalid JSON received"
-        else:
-            return valid, response
